+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes for the solr-cmd-utils project will be documented in this fi
 - implement a filter to add static values
 - implement a random iban generator
 - implement a QSF Data reader
+- implement a qsf data repository manager
 
 ### Changed